requirements.txt
-e .
aws-cdk.aws-ec2==1.81.0
aws-cdk.aws-ecs==1.81.0
aws_cdk.aws_ecs_patterns==1.81.0